refactor: move PageLinkElement to standalone file
