+++ apps/studio/src/components/PageEditor/LinkEditorModal.tsx
@@ -8,7 +8,6 @@ import {
   ModalFooter,
   ModalHeader,
   ModalOverlay,
-  Text,
 } from "@chakra-ui/react"
 import {
   Button,
@@ -24,45 +23,8 @@ import { LinkHrefEditor } from "~/features/editing-experience/components/LinkEdi
 import { editPageSchema } from "~/features/editing-experience/schema"
 import { useQueryParse } from "~/hooks/useQueryParse"
 import { useZodForm } from "~/lib/form"
-import { getReferenceLink, getResourceIdFromReferenceLink } from "~/utils/link"
-import { trpc } from "~/utils/trpc"
-import { ResourceSelector } from "../ResourceSelector"
 import { FileAttachment } from "./FileAttachment"
-
-interface PageLinkElementProps {
-  value: string
-  onChange: (value: string) => void
-}
-
-const PageLinkElement = ({ value, onChange }: PageLinkElementProps) => {
-  const { siteId } = useQueryParse(editPageSchema)
-
-  const selectedResourceId = getResourceIdFromReferenceLink(value)
-
-  const { data: resource } = trpc.resource.getWithFullPermalink.useQuery({
-    resourceId: selectedResourceId,
-  })
-
-  return (
-    <>
-      <ResourceSelector
-        siteId={String(siteId)}
-        onChange={(resourceId) =>
-          onChange(getReferenceLink({ siteId: String(siteId), resourceId }))
-        }
-        selectedResourceId={selectedResourceId}
-      />
-
-      {!!resource && (
-        <Box bg="utility.feedback.info-subtle" p="0.75rem" w="full" mt="0.5rem">
-          <Text textStyle="caption-1">
-            You selected /{resource.fullPermalink}
-          </Text>
-        </Box>
-      )}
-    </>
-  )
-}
+import { PageLinkElement } from "./PageLinkElement"
 
 interface LinkEditorModalContentProps {
   linkText?: string
